feat: add support for UsePathStyle in ServiceConfig
This commit adds the `UsePathStyle` field to the `ServiceConfig` struct in `pkg/config/service.go`. The `UsePathStyle` parameter is optional and allows users to specify whether to use path-style URLs for S3 requests. The value of `UsePathStyle` is read from the `USE_PATH_STYLE` environment variable and defaults to `false`.

Refactor the `CreateClientWithCustomEndpoint` function in `pkg/s3/s3.go` to set the `UsePathStyle` option based on the value of `svcConf.UsePathStyle`. This ensures that the S3 client uses path-style URLs when the `UsePathStyle` parameter is enabled.

Signed-off-by: Cedric Grard <cedric.grard@eurofiber.com>
Signed-off-by: Noel Georgi <git@frezbo.dev>