feat: add support for encryption as optional in BackupSnapshot
This commit modifies the `BackupSnapshot` function in the `service` package to add support for encryption of the etcd snapshot before uploading it to S3. The `BackupSnapshot` function now takes an additional boolean parameter `encrypt` which determines whether encryption should be enabled or not. If encryption is enabled, the etcd snapshot is encrypted using the provided public key before uploading it to S3. If encryption is disabled, the etcd snapshot is uploaded as is. This change allows users to choose whether they want to encrypt their etcd snapshots or not.

Signed-off-by: Cedric Grard <cedric.grard@eurofiber.com>
Signed-off-by: Noel Georgi <git@frezbo.dev>